LMS: WiFiProfileSyncService: replace enums with constants
There is no need for enums AuthenticationMethods and EncryptionMethods.
Replace them with typed constants.

+++ UNS/WiFiProfileSyncService/WlanBL.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /*
- * Copyright (C) 2018-2020 Intel Corporation
+ * Copyright (C) 2018-2022 Intel Corporation
  */
 #ifndef __WLAN_BL_H_
 #define __WLAN_BL_H_
@@ -21,7 +21,7 @@ namespace wlanps {
 		static std::mutex _updateMutex;
 
 		static bool trans2CIM(PINTEL_PROFILE_DATA profileData, SingleMeProfile& wifiSettings);
-		static void PrintWifiSetting(int auth, int enc, int prio, wchar_t* elementName, wchar_t* ssid);
+		static void PrintWifiSetting(unsigned short auth, unsigned short enc, int prio, wchar_t* elementName, wchar_t* ssid);
 		static void CleanOsProfileList(WlanOsProfileList &wlanOsProfiles);
 		static bool AddMissingProfilesToMe(WlanWSManClient &wsmanClient, MeProfileList &MeProfileList, WlanOsProfileList &wlanOsProfiles);
 		static bool FetchOsProfiles(HANDLE hwlan, WlanOsProfileList &wlanOsProfiles);
